+++ src/main/java/com/cqse/teamscaleupload/TeamscaleUpload.java
@@ -16,6 +16,9 @@ import java.net.ConnectException;
 import java.net.UnknownHostException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.time.Instant;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
@@ -48,7 +51,6 @@ public class TeamscaleUpload {
         private final String partition;
         private final String format;
         private final String commit;
-        private final boolean autoDetectCommit;
         private final String timestamp;
         private final HttpUrl url;
         private final List<String> files;
@@ -64,7 +66,6 @@ public class TeamscaleUpload {
             this.accessKey = namespace.getString("accesskey");
             this.partition = namespace.getString("partition");
             this.commit = namespace.getString("commit");
-            this.autoDetectCommit = namespace.getBoolean("detect_commit");
             this.timestamp = namespace.getString("branch_and_timestamp");
             this.files = getListSafe(namespace, "files");
             this.url = HttpUrl.parse(namespace.getString("server"));
@@ -131,8 +132,8 @@ public class TeamscaleUpload {
             }
 
             if (hasMoreThanOneCommitOptionSet()) {
-                throw new ArgumentParserException("You used more than one of --detect-commit, --commit and --timestamp." +
-                        " You must choose one of these three options to specify the commit for which you would like to" +
+                throw new ArgumentParserException("You used more than one of --commit and --branch-and-timestamp." +
+                        " You must choose one of these options to specify the commit for which you would like to" +
                         " upload data to Teamscale", parser);
             }
 
@@ -143,19 +144,60 @@ public class TeamscaleUpload {
             }
 
             if (!files.isEmpty() && format == null) {
-                throw new ArgumentParserException("Please specify a report format with --format " +
-                        "if you pass report patterns as command line arguments", parser);
+                throw new ArgumentParserException("Please specify a report format with --format" +
+                        " if you pass report patterns as command line arguments", parser);
+            }
+
+            validateBranchAndTimestamp(parser);
+        }
+
+        private void validateBranchAndTimestamp(ArgumentParser parser) throws ArgumentParserException {
+            if (timestamp == null) {
+                return;
+            }
+
+            String[] parts = timestamp.split(":", 2);
+            if (parts.length == 1) {
+                throw new ArgumentParserException("You specified an invalid branch and timestamp" +
+                        " with --branch-and-timestamp: " + timestamp + "\nYou must  use the" +
+                        " format BRANCH:TIMESTAMP, where TIMESTAMP is a Unix timestamp in milliseconds" +
+                        " or the string 'HEAD' (to upload to the latest commit on that branch).", parser);
+            }
+
+
+            String timestampPart = parts[1];
+            if (timestampPart.equalsIgnoreCase("HEAD")) {
+                return;
+            }
+
+            validateTimestamp(parser, timestampPart);
+        }
+
+        private void validateTimestamp(ArgumentParser parser, String timestampPart) throws ArgumentParserException {
+            try {
+                long unixTimestamp = Long.parseLong(timestampPart);
+                if (unixTimestamp < 10000000000L) {
+                    String millisecondDate = DateTimeFormatter.RFC_1123_DATE_TIME.format(
+                            Instant.ofEpochMilli(unixTimestamp).atZone(ZoneOffset.UTC));
+                    String secondDate = DateTimeFormatter.RFC_1123_DATE_TIME.format(
+                            Instant.ofEpochSecond(unixTimestamp).atZone(ZoneOffset.UTC));
+                    throw new ArgumentParserException("You specified an invalid timestamp with" +
+                            " --branch-and-timestamp. The timestamp '" + timestampPart + "'" +
+                            " is equal to " + millisecondDate + ". This is probably not what" +
+                            " you intended. Most likely you specified the timestamp in seconds," +
+                            " instead of milliseconds. If you use " + timestampPart + "000" +
+                            " instead, it will mean " + secondDate, parser);
+                }
+            } catch (NumberFormatException e) {
+                throw new ArgumentParserException("You specified an invalid timestamp with" +
+                        " --branch-and-timestamp. Expected either 'HEAD' or a unix timestamp" +
+                        " in milliseconds since 00:00:00 UTC Thursday, 1 January 1970, e.g." +
+                        " master:1606743774000\nInstead you used: " + timestampPart, parser);
             }
         }
 
         private boolean hasMoreThanOneCommitOptionSet() {
-            if (commit != null && timestamp != null) {
-                return true;
-            }
-            if (commit != null && autoDetectCommit) {
-                return true;
-            }
-            return timestamp != null && autoDetectCommit;
+            return commit != null && timestamp != null;
         }
     }
 
@@ -218,11 +260,6 @@ public class TeamscaleUpload {
         parser.addArgument("-i", "--input").metavar("INPUT").required(false)
                 .help("A file which contains additional report file patterns. See INPUTFILE for a" +
                         " detailed description of the file format.");
-        parser.addArgument("--detect-commit").action(Arguments.storeTrue()).required(false)
-                .help("Tries to automatically detect the code commit to which to upload from" +
-                        " environment variables or a Git or SVN checkout in the current working" +
-                        " directory. If guessing fails, the upload will fail. This feature supports" +
-                        " many common CI tools like Jenkins, GitLab, GitHub Actions, Travis CI etc.");
         parser.addArgument("--validate-ssl").action(Arguments.storeTrue()).required(false)
                 .help("By default, SSL certificates are accepted without validation, which makes" +
                         " using this tool with self-signed certificates easier. This flag enables" +
@@ -246,6 +283,14 @@ public class TeamscaleUpload {
         parser.epilog("For general usage help and alternative upload methods, please check our online" +
                 " documentation at:" +
                 "\nhttp://cqse.eu/tsu-docs" +
+                "\n\nTARGET COMMIT" +
+                "\n\nBy default, teamscale-upload tries to automatically detect the code commit" +
+                " to which to upload from environment variables or a Git or SVN checkout in the" +
+                " current working directory. If guessing fails, the upload will fail. This feature" +
+                " supports many common CI tools like Jenkins, GitLab, GitHub Actions, Travis CI etc." +
+                " If automatic detection fails, you can manually specify either a commit via --commit," +
+                " a branch and timestamp via --branch-and-timestamp or you can upload to the latest" +
+                " commit on a branch via --branch-and-timestamp my-branch:HEAD." +
                 "\n\nINPUTFILE" +
                 "\n\nThe input file allows to upload multiple report files for different formats in one" +
                 " upload session. Each section of reports must start with a specification of the" +
@@ -385,16 +430,14 @@ public class TeamscaleUpload {
         } else if (input.timestamp != null) {
             builder.addQueryParameter("t", input.timestamp);
             return input.timestamp;
-        } else if (input.autoDetectCommit) {
+        } else {
+            // auto-detect if neither option is given
             String commit = detectCommit();
             if (commit == null) {
                 fail("Failed to automatically detect the commit. Please specify it manually via --commit or --branch-and-timestamp");
             }
             builder.addQueryParameter("revision", commit);
             return commit;
-        } else {
-            builder.addQueryParameter("t", "HEAD");
-            return "HEAD";
         }
     }