+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We use [semantic versioning](http://semver.org/):
 
 # Next Release
 
+- [breaking change] remove `--detect-commit` option and make that the default behaviour
 - [feature] allow users to provide self-signed certificates that should be trusted in a Java keystore
 - [feature] allow users to append lines to the default message